# openapi/pinuyumayan-main-site-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pinuyumayan Main Site API v25 - Deployable Main Site Bridge + Review/HMAC/SEO Governance",
    "version": "v28",
...
  },
  "servers": [
    {
      "url": "https://kb-api.pinuyumayan.tw"
    },
    {
      "url": "http://localhost:8787"
    }
  ],
...
  "paths": {
...
    "/api/internal/corpus/v27/acceptance-report": {
      "post": {
        "summary": "Record full corpus acceptance report from VPS staging",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "MainSiteHmac": []
          }
        ]
      }
    },
...
    "/api/internal/search/v27/rebuild": {
      "post": {
        "summary": "Queue v27 search index rebuild",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "MainSiteHmac": []
          }
        ]
      }
    },
...
    "/api/internal/vps-db/v27/restore-drill-report": {
      "post": {
        "summary": "Record VPS backup restore drill report",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "MainSiteHmac": []
          }
        ]
      }
    },
...
    "/api/internal/corpus/v28/acceptance-report": {
      "post": {
        "summary": "Submit full corpus acceptance report",
        "tags": [
          "v28 VPS Live Ops"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "MainSiteApiKey": []
          }
        ]
      }
    },
...
    "/api/internal/search/v28/rebuild": {
      "post": {
        "summary": "Queue or dry-run search index rebuild",
        "tags": [
          "v28 VPS Live Ops"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "MainSiteApiKey": []
          }
        ]
      }
    },
...
    "/api/internal/vps-db/v28/restore-drill-report": {
      "post": {
        "summary": "Submit VPS restore drill report",
        "tags": [
          "v28 VPS Live Ops"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "MainSiteApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "mainSiteApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pinuyumayan-main-site-key"
      },
      "MainSiteApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pinuyumayan-main-site-key"
      },
      "MainSiteHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pinuyumayan-signature"
      }
    },
...
  }
}